+++ package.json
@@ -53,7 +53,7 @@
     "prettier": "^3.0.1",
     "storybook": "^7.2.0",
     "typescript": "^5.0.2",
-    "vite": "^4.4.5"
+    "vite": "^4.4.9"
   },
   "packageManager": "yarn@3.6.1"
 }
