+++ src/App.tsx
@@ -1,43 +1,13 @@
 import './App.css'
 import { TamaguiProvider } from 'tamagui'
 import { Provider as StatusProvider } from '@status-im/components'
-import { ResponsivePie } from '@nivo/pie'
-
+import StandartGauge from './components/StandartGauge'
+ 
 import config from '../tamagui.config'
 
-interface Data {
-  id: string
-  label: string
-  value: number
-  color: string
-}
-interface StandartGaugeProps {
-  data: Data[]
-}
-
-const StandartGauge = (props: StandartGaugeProps) => {
-  const { data } = props
-
-  return (
-    <>
-      <ResponsivePie
-        data={data}
-        margin={{ top: 40, right: 80, bottom: 80, left: 80 }}
-        innerRadius={0.65}
-        colors={datum => datum.data.color}
-        fit={false}
-        activeOuterRadiusOffset={8}
-        enableArcLinkLabels={false}
-        arcLinkLabelsColor={{ from: 'color' }}
-        enableArcLabels={false}
-        legends={[]}
-      />
-    </>
-  )
-}
 
 function App() {
-  const data: Data[] = [
+  const data = [
     {
       id: 'php',
       label: 'php',
